Extract feed ingestion from NewsModels into a service
Move featuredBuilder and its feed-fetching pipeline (getFeaturedFeeds,
processSingleFeed, saveFeedItems and the deadlock-retry helpers) out of
NewsModels into a new FeedIngestionService, leaving the model with data
access only. This brings NewsModels below the PHPMD ExcessiveClassComplexity
threshold. The featuredBuilder and deadlock-retry tests move to
FeedIngestionServiceTest accordingly.
